Complete namespace changes. Add Trevor's new weighted extents and other updates. Ready for merge into trunk.

+++ indri/branches/dmf-devel/include/indri/DirichletTermScoreFunction.hpp
@@ -39,12 +39,12 @@ namespace indri
 	_muTimesCollectionFrequency = _mu * _collectionFrequency;
       }
 
-      double scoreOccurrence( int occurrences, int contextSize ) {
+      double scoreOccurrence( double occurrences, int contextSize ) {
 	double seen = ( double(occurrences) + _muTimesCollectionFrequency ) / ( double(contextSize) + _mu );
 	return log( seen );
       }
 
-      double scoreOccurrence( int occurrences, int contextSize, int documentOccurrences, int documentLength ) {
+      double scoreOccurrence( double occurrences, int contextSize, double documentOccurrences, int documentLength ) {
 	// can't two-level smooth with dirichlet
 	return scoreOccurrence( occurrences, contextSize );
       }
