+++ indri/branches/dmf-devel/include/indri/TermScoreFunction.hpp
@@ -31,8 +31,8 @@ namespace indri
     */
     class TermScoreFunction {
     public:
-      virtual double scoreOccurrence( int occurrences, int contextLength ) = 0;
-      virtual double scoreOccurrence( int occurrences, int contextLength, int documentOccurrences, int documentLength ) = 0;
+      virtual double scoreOccurrence( double occurrences, int contextLength ) = 0;
+      virtual double scoreOccurrence( double occurrences, int contextLength, double documentOccurrences, int documentLength ) = 0;
     };
   }
 }
